input/fp||ts: move new sensor attrs to u8 from int
Inspired from
https://github.com/SOVIET-ANDROID/kernel_xiaomi_raphael/commit/7991b6efdfa238ea10c1115c28ded261ecabf057

+++ drivers/input/oneplus_touchscreen/touchpanel_common.h
@@ -405,7 +405,7 @@ struct touchpanel_data {
 
 	void *chip_data;	/*Chip Related data */
 	void *private_data;	/*Reserved Private data */
-	int double_tap_pressed;
+	u8 double_tap_pressed;
 };
 
 struct touchpanel_operations {
